+++ test/test-slots.js
@@ -75,7 +75,7 @@ test('Must process slots conditional rendering by using slot name', async t => {
   t.is(html, expected);
 });
 
-test('Must render slots using $slots locals', async t => {
+test('Must render slots using $slots props', async t => {
   const actual = `<component src="layouts/base-render-slots-locals.html"><div>Main content</div><fill:header><h1>My header</h1></fill:header><fill:footer>My footer</fill:footer></component>`;
   const expected = `<html><head><title>Base Render Slots Locals Layout</title></head><body><header><h1>My header</h1></header><main><div>Main content</div></main><footer>My footer</footer></body></html>`;
 
@@ -84,7 +84,7 @@ test('Must render slots using $slots locals', async t => {
   t.is(html, expected);
 });
 
-test('Must merge global locals passed via expressions.locals', async t => {
+test('Must merge global props passed via expressions.locals', async t => {
   const actual = `<component src="layouts/global-locals.html" anArray='["third", "fourth"]'><div>Main content</div></component>`;
   const expected = `<html><head><title>My page title</title></head><body><main><div>Main content</div></main><span>first</span><span>second</span><span>third</span><span>fourth</span></body></html>`;